Приложение для предзащиты.

+++ BagProduct+CoreDataProperties.swift
@@ -23,7 +23,7 @@ extension BagProduct {
     @NSManaged public var nameProduct: String?
     @NSManaged public var count: Int16
     @NSManaged public var category: String?
-    @NSManaged public var price: Int16
+    @NSManaged public var price: Int64
 }
 
 extension BagProduct : Identifiable {
