+++ BonchTech Store/Core/CodeDataManager.swift
@@ -26,13 +26,13 @@ public final class CoreDataManager: NSObject{
 //        }
 //    }
     
-    public func createProduct(name nameProduct: String, category: String, price: Int16 ) {
+    public func createProduct(name nameProduct: String, category: String, price: Int64 ) {
         guard let productEntityDescription = NSEntityDescription.entity(forEntityName: "BagProduct", in: context) else { return }
         let product = BagProduct(entity: productEntityDescription, insertInto: context)
         product.nameProduct = nameProduct
         product.count = 1
         product.category = category
-        product.price = price
+        product.price = Int64(price)
         
         appDelegate.saveContext()
     }
